main implementation
